260522/안형주/Feat/그룹에 포함된 프로젝트만 출력(회원),프로젝트 생성프로시저 적용중,프로젝트 메인 출력완료

+++ FinalProject/src/main/java/com/example/demo/project/notice/controller/NoticeController.java
@@ -59,7 +59,7 @@ public class NoticeController {
     //목록페이지
     @GetMapping("/notice/list")
     public String noticeList(Model model,NoticeVO notice) {
-    	notice.setPrjId(4);
+    	notice.setPrjId((long) 4);
         model.addAttribute("list", noticeService.selectAll(notice));
         return "project/notice/noticeList";
     }
